Reference

Contents

Reference

Contents

Index

Index

ChemistryLab.issubscript

ChemistryLab.issuperscript

ChemistryLab.normal_to_sub

ChemistryLab.normal_to_super

ChemistryLab.sub_to_normal

ChemistryLab.super_to_normal

ChemistryLab.issubscript — Method

Return whether c is a numeric subscript.

ChemistryLab.issuperscript — Method

Return whether c is a numeric superscript or ⁺/⁻.

ChemistryLab.normal_to_sub — Method

Convert all normal characters in s to numeric subscripts .

ChemistryLab.normal_to_super — Method

Convert all normal characters or +/- in s to numeric superscripts.

ChemistryLab.sub_to_normal — Method

Convert all numeric subscripts in s to normal line.

ChemistryLab.super_to_normal — Method

Convert all numeric superscripts or ⁺/⁻ in s to normal line.